+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Release NEXT:
  - Change several integer properties to durations, e.g. expirations (in seconds) for OIDC data classes
  - In `SupportedCredentialFormat` replace `claims` with `isoClaims` and `sdJwtClaims` to be able to handle both formats defined in OID4VCI Draft 13
  - Wrap exceptions during deserialization in `KmmResult`, i.e. changing all `deserialize()` methods in companion objects
+ - `OidcSiopWallet`: Rename `newInstance()` to `newDefaultInstance()`, to align it with other factory methods
+ - `OidcSiopWallet`: Rename `retrieveAuthenticationRequestParameters() ` to `parseAuthenticationRequestParameters()`, changing result type to `KmmResult<AuthenticationRequestParameters>`
 
 Release 3.6.1:
  * Update to KMP-Crypto 2.6.0